Correct logical errors
