+++ src/index.js
@@ -1,6 +1,5 @@
 import { Provider } from 'react-redux';
 import App from './App';
-import reportWebVitals from './reportWebVitals';
 import store from './redux/configureStore';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
@@ -11,4 +10,3 @@ root.render(
     </Provider>
   </React.StrictMode>,
 );
-reportWebVitals();
